cifs: rename module to smbfs.ko
This commit is the start of an effort to move away from the old "CIFS"
nomenclature, which is associated with the old, insecure SMBv1.

For now, only rename the module to "smbfs.ko" and update Kconfigs to
reference it as "CONFIG_SMBFS".

Keep "cifs" as a module alias for backward compatibility.

This patch doesn't touch any functional code, aside from changing
references to the module to the new name.

Signed-off-by: Enzo Matsumiya <ematsumiya@suse.de>