smbfs: major refactoring in debug/logging
This commit changes the whole debugging and logging infrastructure for
smbfs module. This is a preparation for the decoupling of SMB1 and SMB2+
code.

This shouldn't change any functional (as in, protocol related)
behaviour, but if some application depended on the old procfs entries,
they will need to be updated.

- split old debug code into procfs-specific code and debugging/logging
  routines
- create a proc.c file for the procfs-specific routines
- standardize debugging/logging message formats
- convert CamelCase to snake_case for related code (WIP for whole module)
- removed no longer used server ops (.print_stats, .clear_stats,
  .dump_detail, etc)
- replace list_for_each() by list_for_each_entry()
- several other minor cosmetic changes

Signed-off-by: Enzo Matsumiya <ematsumiya@suse.de>