cifs: remove "cifs_" prefix from SMB2+ related files
Since all SMB1 (CIFS) files are now inside "smb1" dir, remove the
"cifs_" prefix from the files in fs/cifs.

The aim is to make this module as version-agnostic as possible. It also
makes it easier to rename the module to something more modern and
version-agnostic as well.

Nodule hierarchy becomes:
- fs/cifs/*.{c,h}: Main module code, code related to SMB2.0 and later,
		   and their relative features.
- fs/cifs/smb1/*.{c,h}: Code specific to SMB1 and its related features.

This prettifies the module structure by making it cleaner, as the scopes
(protocol versions) are defined by separated directories.

Signed-off-by: Enzo Matsumiya <ematsumiya@suse.de>